requirements.txt
fastmcp
streamlit
crewai
crewai-tools
openai
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sentence-transformers
faiss-cpu
pandas
beautifulsoup4
lxml

